+++ crates/core/src/task_center/task_kind.rs
@@ -8,15 +8,11 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use futures::FutureExt;
-use std::pin::Pin;
-use std::task::{ready, Context, Poll};
-use strum::EnumProperty;
-use tokio::runtime::RuntimeMetrics;
-use tokio::sync::oneshot;
-use tokio_util::sync::CancellationToken;
+use std::sync::atomic::{AtomicU64, Ordering};
 
-use crate::ShutdownError;
+use strum::EnumProperty;
+
+static NEXT_TASK_ID: AtomicU64 = const { AtomicU64::new(0) };
 
 #[derive(
     Clone,
@@ -32,6 +28,18 @@ use crate::ShutdownError;
     derive_more::Into,
 )]
 pub struct TaskId(u64);
+
+impl Default for TaskId {
+    fn default() -> Self {
+        Self(NEXT_TASK_ID.fetch_add(1, Ordering::Relaxed))
+    }
+}
+
+impl TaskId {
+    pub fn new() -> Self {
+        Default::default()
+    }
+}
 
 /// Describes the types of tasks TaskCenter manages.
 ///
@@ -168,100 +176,4 @@ pub enum AsyncRuntime {
     Default,
     /// Run on ingress runtime
     Ingress,
-}
-
-/// A future that represents a task spawned on the TaskCenter.
-///
-/// Awaiting this future waits for the task to complete.
-pub struct TaskHandle<T> {
-    pub(crate) cancellation_token: CancellationToken,
-    pub(crate) inner_handle: tokio::task::JoinHandle<T>,
-}
-
-impl<T> TaskHandle<T> {
-    /// Abort the task immediately. This will abort the task at the next yielding point. If the
-    /// task is running a blocking call, it'll not be aborted until it can yield to the runtime.
-    pub fn abort(&self) {
-        self.inner_handle.abort();
-    }
-
-    /// Trigger graceful cancellation of the task
-    pub fn cancel(&self) {
-        self.cancellation_token.cancel()
-    }
-
-    /// Returns true if cancellation was requested. Note that this doesn't mean that
-    /// the task has finished. To check if the task has finished or not, use `is_finished()`
-    pub fn is_cancellation_requested(&self) -> bool {
-        self.cancellation_token.is_cancelled()
-    }
-
-    /// Returns true if the task has finished executing. Note that this might return
-    /// `false` after calling `abort()` since termination process takes time.
-    pub fn is_finished(&self) -> bool {
-        self.inner_handle.is_finished()
-    }
-}
-
-impl<T> std::future::Future for TaskHandle<T> {
-    type Output = Result<T, ShutdownError>;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut std::task::Context<'_>) -> Poll<Self::Output> {
-        match ready!(self.inner_handle.poll_unpin(cx)) {
-            Ok(v) => Poll::Ready(Ok(v)),
-            Err(_) => Poll::Ready(Err(ShutdownError)),
-        }
-    }
-}
-
-/// A handle for a dedicated runtime managed by task-center
-pub struct RuntimeHandle {
-    pub(crate) cancellation_token: CancellationToken,
-    pub(crate) inner: tokio::runtime::Runtime,
-}
-
-impl RuntimeHandle {
-    /// Trigger graceful shutdown of the runtime. This will trigger the cancellation token of the
-    /// root localset running on this runtime. Shutdown is not guaranteed, it depends on whether
-    /// the root localset awaits the cancellation token or not.
-    pub fn cancel(&self) {
-        self.cancellation_token.cancel()
-    }
-
-    pub fn metrics(&self) -> RuntimeMetrics {
-        self.inner.metrics()
-    }
-
-    /// Returns true if cancellation was requested. Note that this doesn't mean that
-    /// the runtime has been terminated.
-    pub fn is_cancellation_requested(&self) -> bool {
-        self.cancellation_token.is_cancelled()
-    }
-}
-
-pub struct RuntimeRootTaskHandle<T> {
-    pub(crate) cancellation_token: CancellationToken,
-    pub(crate) inner_handle: oneshot::Receiver<T>,
-}
-
-impl<T> RuntimeRootTaskHandle<T> {
-    /// Trigger graceful shutdown of the runtime root task. Shutdown is not guaranteed, it depends
-    /// on whether the root task awaits the cancellation token or not.
-    pub fn cancel(&self) {
-        self.cancellation_token.cancel()
-    }
-
-    pub fn cancellation_token(&self) -> CancellationToken {
-        self.cancellation_token.clone()
-    }
-}
-
-impl<T> std::future::Future for RuntimeRootTaskHandle<T> {
-    type Output = T;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        Poll::Ready(
-            ready!(self.inner_handle.poll_unpin(cx)).expect("runtime panicked unexpectedly"),
-        )
-    }
 }
